Restructure modules so that it will make sense to have a second data source
* Rename `QueryData` to `FileSystemDataSource`, to contrast with
  `SqlDataSource` which we will be adding soon
* Move `FileSystemDataSource` to its own module under `data_source`,
  so we'll have `data_source::fs` and later `data_source::sql`
* Move some data source related types (LedgerLog and UpdateDataSource)
  under the `data_source` module
* Make data source tests generic and instantiate for `FileSystemDataSource`
* Restructure documentation, so crate-level documentation is now largely
  data source agnostic, and refers to the documentation of the individual
  data sources for data-source-specific things like extension and
  composability

Closes #213
