+++ src/availability.rs
@@ -286,9 +286,9 @@ where
 mod test {
     use super::*;
     use crate::{
-        data_source::QueryData,
+        data_source::FileSystemDataSource,
         testing::{
-            consensus::MockNetwork,
+            consensus::{MockDataSource, MockNetwork},
             mocks::{MockNodeImpl, MockTransaction, MockTypes},
             setup_test,
         },
@@ -464,7 +464,7 @@ mod test {
         setup_test();
 
         // Create the consensus network.
-        let mut network = MockNetwork::init(()).await;
+        let mut network = MockNetwork::<MockDataSource>::init().await;
         network.start().await;
 
         // Start the web server.
@@ -524,7 +524,8 @@ mod test {
         setup_test();
 
         let dir = TempDir::new("test_availability_extensions").unwrap();
-        let query_data = QueryData::<MockTypes, MockNodeImpl, u64>::create(dir.path(), 0).unwrap();
+        let query_data =
+            FileSystemDataSource::<MockTypes, MockNodeImpl, u64>::create(dir.path(), 0).unwrap();
 
         // Create the API extensions specification.
         let extensions = toml! {
@@ -538,14 +539,15 @@ mod test {
             METHOD = "GET"
         };
 
-        let mut api =
-            define_api::<RwLock<QueryData<MockTypes, MockNodeImpl, u64>>, MockTypes, MockNodeImpl>(
-                &Options {
-                    extensions: vec![extensions.into()],
-                    ..Default::default()
-                },
-            )
-            .unwrap();
+        let mut api = define_api::<
+            RwLock<FileSystemDataSource<MockTypes, MockNodeImpl, u64>>,
+            MockTypes,
+            MockNodeImpl,
+        >(&Options {
+            extensions: vec![extensions.into()],
+            ..Default::default()
+        })
+        .unwrap();
         api.get("get_ext", |_, state| {
             async move { Ok(*state.as_ref()) }.boxed()
         })
