+++ src/availability/data_source.rs
@@ -83,14 +83,14 @@ pub trait AvailabilityDataSource<Types: NodeType, I: NodeImplementation<Types>>
 where
     Block<Types>: QueryableBlock,
 {
-    type LeafStream: Stream<Item = LeafQueryData<Types, I>> + Send;
-    type BlockStream: Stream<Item = BlockQueryData<Types>> + Send;
+    type LeafStream: Stream<Item = LeafQueryData<Types, I>> + Unpin + Send;
+    type BlockStream: Stream<Item = BlockQueryData<Types>> + Unpin + Send;
 
-    type LeafRange<'a, R>: 'a + Stream<Item = QueryResult<LeafQueryData<Types, I>>>
+    type LeafRange<'a, R>: 'a + Stream<Item = QueryResult<LeafQueryData<Types, I>>> + Unpin
     where
         Self: 'a,
         R: RangeBounds<usize> + Send;
-    type BlockRange<'a, R>: 'a + Stream<Item = QueryResult<BlockQueryData<Types>>>
+    type BlockRange<'a, R>: 'a + Stream<Item = QueryResult<BlockQueryData<Types>>> + Unpin
     where
         Self: 'a,
         R: RangeBounds<usize> + Send;
