+++ src/lib.rs
@@ -28,7 +28,7 @@
 //! # async fn doc(storage_path: &std::path::Path) -> Result<(), hotshot_query_service::Error> {
 //! use hotshot_query_service::{
 //!     availability,
-//!     data_source::{UpdateDataSource, QueryData},
+//!     data_source::{FileSystemDataSource, UpdateDataSource},
 //!     status, Error
 //! };
 //!
@@ -37,8 +37,8 @@
 //! use hotshot::SystemContext;
 //! use tide_disco::App;
 //!
-//! // Create or open query data.
-//! let query_data = QueryData::<AppTypes, AppNodeImpl, ()>::create(storage_path, ())
+//! // Create or open a data source.
+//! let query_data = FileSystemDataSource::<AppTypes, AppNodeImpl, ()>::create(storage_path, ())
 //!     .map_err(Error::internal)?;
 //!
 //! // Create hotshot, giving it a handle to the status metrics.
@@ -86,17 +86,24 @@
 //! ```
 //! # use async_std::task::spawn;
 //! # use hotshot::types::SystemContextHandle;
-//! # use hotshot_query_service::{data_source::QueryData, Error, Options};
+//! # use hotshot_query_service::{data_source::FileSystemDataSource, Error, Options};
 //! # use hotshot_query_service::testing::mocks::{MockTypes, MockNodeImpl};
 //! # use std::path::Path;
 //! # fn doc(storage_path: &Path, options: &Options, hotshot: SystemContextHandle<MockTypes, MockNodeImpl>) -> Result<(), Error> {
 //! use hotshot_query_service::run_standalone_service;
 //!
-//! let query_data = QueryData::create(storage_path, ()).map_err(Error::internal)?;
+//! let query_data = FileSystemDataSource::create(storage_path, ()).map_err(Error::internal)?;
 //! spawn(run_standalone_service(options, query_data, hotshot));
 //! # Ok(())
 //! # }
 //! ```
+//!
+//! # Persistence
+//!
+//! Naturally, an archival query service such as this is heavily dependent on a persistent storage
+//! implementation. The APIs provided by this query service are generic over the specific type of
+//! the persistence layer, which we call a _data source_. This crate provides several data source
+//! implementations in the [`data_source`] module.
 //!
 //! # Interaction with other components
 //!
@@ -125,71 +132,66 @@
 //! about the internal structure of your transactions.
 //!
 //! There are two parts to adding additional functionality to a module in this crate: adding the
-//! required additional data structures to the module state, and creating a new API endpoint to
-//! expose the functionality. For the former, you can take advantage of the `UserData` type
-//! parameter of the [QueryData] state to inject whatever additional data you need. In the case of
-//! adding a UTXO index, it might look like this:
+//! required additional data structures to the data source, and creating a new API endpoint to
+//! expose the functionality. The mechanism for the former will depend on the specific data source
+//! you are using. Check the documentation for your data source implementation to see how it can be
+//! extended.
+//!
+//! For the latter, you can modify the default availablity API with the addition of a new endpoint
+//! that accesses the custom state you have added to the data source. It is good practice to define
+//! a trait for accessing this custom state, so that if you want to switch data sources in the
+//! future, you can easily extend the new data source, implement the trait, and then transparently
+//! replace the data source that you use to set up your API. In the case of
+//! adding a UTXO index, this trait might look like this:
 //!
 //! ```
-//! # use hotshot_query_service::availability::{
-//! #     self, TransactionIndex,
+//! # use hotshot_query_service::{
+//! #   availability::{AvailabilityDataSource, TransactionIndex},
+//! #   testing::mocks::{MockNodeImpl as AppNodeImpl, MockTypes as AppTypes},
 //! # };
-//! # use hotshot_query_service::data_source::QueryData;
-//! # use hotshot_query_service::testing::mocks::{
-//! #   MockNodeImpl as AppNodeImpl, MockTypes as AppTypes,
-//! # };
-//! # use std::collections::HashMap;
-//! #[derive(Default)]
-//! struct AppQueryData {
+//! use async_trait::async_trait;
+//!
+//! #[async_trait]
+//! trait UtxoDataSource: AvailabilityDataSource<AppTypes, AppNodeImpl> {
 //!     // Index mapping UTXO index to (block index, transaction index, output index)
-//!     utxo_index: HashMap<u64, (u64, TransactionIndex<AppTypes>, u64)>,
+//!     async fn find_utxo(&self, utxo: u64) -> Option<(usize, TransactionIndex<AppTypes>, usize)>;
 //! }
-//!
-//! type AvailabilityState = QueryData<AppTypes, AppNodeImpl, AppQueryData>;
 //! ```
 //!
-//! `QueryData<AppTypes, AppNodeImpl, AppQueryData>` implements `AsRef<AppQueryData>` and
-//! `AsMut<AppQueryData>`, so you can now modify the default availablity API with the addition of a
-//! new endpoint that accesses `AppQueryData` like so:
+//! Implement this trait for the extended data source you're using, and then add a new endpoint to
+//! the availability API like so:
 //!
 //! ```
 //! # use async_std::sync::RwLock;
+//! # use async_trait::async_trait;
 //! # use futures::FutureExt;
 //! # use hotshot_query_service::availability::{
 //! #   self, AvailabilityDataSource, QueryBlockSnafu, TransactionIndex,
 //! # };
-//! # use hotshot_query_service::data_source::QueryData;
 //! # use hotshot_query_service::testing::mocks::{
 //! #   MockNodeImpl as AppNodeImpl, MockTypes as AppTypes,
 //! # };
 //! # use hotshot_query_service::Error;
 //! # use snafu::ResultExt;
-//! # use std::collections::HashMap;
-//! # use std::path::Path;
 //! # use tide_disco::{api::ApiError, method::ReadState, Api, App, StatusCode};
-//! #[derive(Default)]
-//! # struct AppQueryData {
-//! #     utxo_index: HashMap<u64, (usize, TransactionIndex<AppTypes>, usize)>,
+//! # #[async_trait]
+//! # trait UtxoDataSource: AvailabilityDataSource<AppTypes, AppNodeImpl> {
+//! #   async fn find_utxo(&self, utxo: u64) -> Option<(usize, TransactionIndex<AppTypes>, usize)>;
 //! # }
-//! # type AvailabilityState = QueryData<AppTypes, AppNodeImpl, AppQueryData>;
+//!
 //! fn define_app_specific_availability_api<State>(
 //!     options: &availability::Options,
 //! ) -> Result<Api<State, availability::Error>, ApiError>
 //! where
 //!     State: 'static + Send + Sync + ReadState,
-//!     <State as ReadState>::State:
-//!         Send +
-//!         Sync +
-//!         AvailabilityDataSource<AppTypes, AppNodeImpl> +
-//!         AsRef<AppQueryData>,
+//!     <State as ReadState>::State: UtxoDataSource + Send + Sync,
 //! {
 //!     let mut api = availability::define_api(options)?;
 //!     api.get("get_utxo", |req, state: &<State as ReadState>::State| async move {
 //!         let utxo_index = req.integer_param("index")?;
-//!         let app_query_data = state.as_ref();
-//!         let (block_index, txn_index, output_index) = *app_query_data
-//!             .utxo_index
-//!             .get(&utxo_index)
+//!         let (block_index, txn_index, output_index) = state
+//!             .find_utxo(utxo_index)
+//!             .await
 //!             .ok_or_else(|| availability::Error::Custom {
 //!                 message: format!("no such UTXO {}", utxo_index),
 //!                 status: StatusCode::NotFound,
@@ -206,15 +208,13 @@
 //!     Ok(api)
 //! }
 //!
-//! fn init_server(
+//! fn init_server<D: UtxoDataSource + Send + Sync + 'static>(
 //!     options: &availability::Options,
-//!     storage_path: &Path,
-//! ) -> Result<App<RwLock<AvailabilityState>, Error>, availability::Error> {
+//!     data_source: D,
+//! ) -> Result<App<RwLock<D>, Error>, availability::Error> {
 //!     let api = define_app_specific_availability_api(options)
 //!         .map_err(availability::Error::internal)?;
-//!     let state = AvailabilityState::create(storage_path, AppQueryData::default())
-//!         .map_err(availability::Error::internal)?;
-//!     let mut app = App::with_state(RwLock::new(state));
+//!     let mut app = App::with_state(RwLock::new(data_source));
 //!     app.register_module("availability", api).map_err(availability::Error::internal)?;
 //!     Ok(app)
 //! }
@@ -240,13 +240,13 @@
 //! you need to do is register the [availability] and [status] APIs provided by this crate with a
 //! [tide_disco::App], and then register your own API modules with the same app.
 //!
-//! The one wrinkle is that all modules within a [tide_disco] app must share the state type. It is
-//! for this reason that the modules provided by this crate are generic on the state type -- both
+//! The one wrinkle is that all modules within a [tide_disco] app must share the same state type. It
+//! is for this reason that the modules provided by this crate are generic on the state type -- both
 //! [availability::define_api] and [status::define_api] can work with any state type, provided that
-//! type implements the corresponding data source traits. The application state provided by this
-//! crate ([data_source::QueryData]) implements both of these traits, but if you want to use a
-//! custom state type that includes state for other modules, you will need to implement these traits
-//! for your custom type. The basic pattern looks like:
+//! type implements the corresponding data source traits. The data sources provided by this crate
+//! implement both of these traits, but if you want to use a custom state type that includes state
+//! for other modules, you will need to implement these traits for your custom type. The basic
+//! pattern looks like this:
 //!
 //! ```
 //! # use async_trait::async_trait;
@@ -255,40 +255,35 @@
 //! #   AvailabilityDataSource, BlockId, BlockQueryData, LeafId, LeafQueryData, QueryResult,
 //! #   TransactionHash, TransactionIndex,
 //! # };
-//! # use hotshot_query_service::data_source::QueryData;
 //! # use hotshot_query_service::status::{MempoolQueryData, StatusDataSource};
 //! # use hotshot_query_service::testing::mocks::{
 //! #   MockNodeImpl as AppNodeImpl, MockTypes as AppTypes,
 //! # };
 //! # use std::ops::RangeBounds;
 //! # type AppQueryData = ();
-//! struct AppState {
-//!     hotshot_qs: QueryData<AppTypes, AppNodeImpl, AppQueryData>,
+//! // Our AppState takes an underlying data source `D` which already implements the relevant
+//! // traits, and adds some state for use with other modules.
+//! struct AppState<D> {
+//!     hotshot_qs: D,
 //!     // additional state for other modules
 //! }
 //!
-//! // Implement data source trait for availability API.
+//! // Implement data source trait for availability API by delegating to the underlying data source.
 //! #[async_trait]
-//! impl AvailabilityDataSource<AppTypes, AppNodeImpl> for AppState {
-//!     type LeafRange<'a, R> =
-//!         <QueryData<AppTypes, AppNodeImpl, AppQueryData> as
-//!             AvailabilityDataSource<AppTypes, AppNodeImpl>>::LeafRange<'a, R>
+//! impl<D: AvailabilityDataSource<AppTypes, AppNodeImpl> + Send + Sync>
+//!     AvailabilityDataSource<AppTypes, AppNodeImpl> for AppState<D>
+//! {
+//!     type LeafRange<'a, R> = D::LeafRange<'a, R>
 //!     where
 //!         Self: 'a,
 //!         R: RangeBounds<usize> + Send;
-//!     type BlockRange<'a, R> =
-//!         <QueryData<AppTypes, AppNodeImpl, AppQueryData> as
-//!             AvailabilityDataSource<AppTypes, AppNodeImpl>>::BlockRange<'a, R>
+//!     type BlockRange<'a, R> = D::BlockRange<'a, R>
 //!     where
 //!         Self: 'a,
 //!         R: RangeBounds<usize> + Send;
 //!
-//!     type LeafStream =
-//!         <QueryData<AppTypes, AppNodeImpl, AppQueryData> as
-//!             AvailabilityDataSource<AppTypes, AppNodeImpl>>::LeafStream;
-//!     type BlockStream =
-//!         <QueryData<AppTypes, AppNodeImpl, AppQueryData> as
-//!             AvailabilityDataSource<AppTypes, AppNodeImpl>>::BlockStream;
+//!     type LeafStream = D::LeafStream;
+//!     type BlockStream = D::BlockStream;
 //!
 //!     async fn get_leaf<ID>(&self, id: ID) -> QueryResult<LeafQueryData<AppTypes, AppNodeImpl>>
 //!     where
@@ -314,10 +309,10 @@
 //! #   async fn subscribe_blocks(&self, height: usize) -> QueryResult<Self::BlockStream> { todo!() }
 //! }
 //!
-//! // Implement data source trait for status API.
+//! // Implement data source trait for status API by delegating to the underlying data source.
 //! #[async_trait]
-//! impl StatusDataSource for AppState {
-//!     type Error = <QueryData<AppTypes, AppNodeImpl, AppQueryData> as StatusDataSource>::Error;
+//! impl<D: StatusDataSource + Send + Sync> StatusDataSource for AppState<D> {
+//!     type Error = D::Error;
 //!
 //!     async fn block_height(&self) -> Result<usize, Self::Error> {
 //!         self.hotshot_qs.block_height().await
@@ -338,98 +333,33 @@
 //! for a custom type that has an existing implementation as one of its fields.
 //!
 //! Once you have created your `AppState` type aggregating the state for each API module, you can
-//! instantiate the state as normal, using [create](QueryData::create) or [open](QueryData::open) to
-//! initialize the [QueryData] part of the `AppState`. _However_, this only works if you want the
-//! persistent storage for the availability and status modules (managed by [QueryData]) to be
-//! independent of the persistent storage for other modules. You may well want to synchronize the
-//! storage for all modules together, so that updates to the entire application state can be done
-//! atomically. This is particularly relevant if one of your application-specific modules updates
-//! its storage based on a stream of HotShot leaves. Since the availability module also updates with
-//! each new leaf, you probably want these two modules to stay in sync.
+//! initialize the state as normal, instantiating `D` with a concrete implementation of a data
+//! source and initializing `hotshot_qs` as you normally would that data source.
 //!
-//! To achieve this, add a top level [AtomicStore](atomic_store::AtomicStore) to `AppState` to
-//! synchronize all persistent storage, and use [create_with_store](QueryData::create_with_store) or
-//! [open_with_store](QueryData::open_with_store) when initializeing the [QueryData] to associate
-//! its persistent storage with the top-level atomic store. In this case, [QueryData] will not
-//! manage the [AtomicStore](atomic_store::AtomicStore) for you, so you must be sure to call
-//! [AtomicStore::commit_version](atomic_store::AtomicStore::commit_version) after each call to
-//! [QueryData::commit_version], and you must call either [QueryData::commit_version] or
-//! [QueryData::skip_version] before each call to
-//! [AtomicStore::commit_version](atomic_store::AtomicStore::commit_version).
-//!
-//! ```
-//! # use async_std::{sync::{Arc, RwLock}, task::spawn};
-//! # use atomic_store::{AtomicStore, AtomicStoreLoader};
-//! # use futures::StreamExt;
-//! # use hotshot::types::SystemContextHandle;
-//! # use hotshot_query_service::Error;
-//! # use hotshot_query_service::data_source::{UpdateDataSource, QueryData};
-//! # use hotshot_query_service::testing::mocks::{
-//! #   MockNodeImpl as AppNodeImpl, MockTypes as AppTypes,
-//! # };
-//! # use std::path::Path;
-//! # use tide_disco::App;
-//! # type AppQueryData = ();
-//! struct AppState {
-//!     // Top-level storage coordinator
-//!     store: AtomicStore,
-//!     hotshot_qs: QueryData<AppTypes, AppNodeImpl, AppQueryData>,
-//!     // additional state for other modules
-//! }
-//!
-//! fn init_server(
-//!     storage_path: &Path,
-//!     mut hotshot: SystemContextHandle<AppTypes, AppNodeImpl>,
-//! ) -> Result<App<Arc<RwLock<AppState>>, Error>, Error> {
-//!     let mut loader = AtomicStoreLoader::create(storage_path, "my_app") // or `open`
-//!         .map_err(Error::internal)?;
-//!     let hotshot_qs = QueryData::create_with_store(&mut loader, AppQueryData::default())
-//!         .map_err(Error::internal)?;
-//!     // Initialize storage for other modules using the same loader.
-//!
-//!     let store = AtomicStore::open(loader).map_err(Error::internal)?;
-//!     let state = Arc::new(RwLock::new(AppState {
-//!         store,
-//!         hotshot_qs,
-//!         // additional state for other modules
-//!     }));
-//!     let mut app = App::with_state(state.clone());
-//!     // Register API modules.
-//!
-//!     spawn(async move {
-//!         let mut events = hotshot.get_event_stream(Default::default()).await.0;
-//!         while let Some(event) = events.next().await {
-//!             let mut state = state.write().await;
-//!             state.hotshot_qs.update(&event).await.unwrap();
-//!             // Update other modules' states based on `event`.
-//!
-//!             state.hotshot_qs.commit_version().await.unwrap();
-//!             // Commit or skip versions for other modules' storage.
-//!             state.store.commit_version().unwrap();
-//!         }
-//!     });
-//!
-//!     Ok(app)
-//! }
-//! ```
+//! _However_, this only works if you want the persistent storage for the availability and status
+//! modules (managed by `hotshot_qs`) to be independent of the persistent storage for other modules.
+//! You may well want to synchronize the storage for all modules together, so that updates to the
+//! entire application state can be done atomically. This is particularly relevant if one of your
+//! application-specific modules updates its storage based on a stream of HotShot leaves. Since the
+//! availability module also updates with each new leaf, you probably want these two modules to stay
+//! in sync. The data source implementations provided by this crate provide means by which you can
+//! add additional data to the same persistent store and synchronize the entire store together.
+//! Refer to the documentation for you specific data source for information on how to achieve this.
 //!
 
 mod api;
 pub mod availability;
 pub mod data_source;
 mod error;
-mod ledger_log;
 mod metrics;
 mod resolvable;
 pub mod status;
 pub mod testing;
-mod update;
 
 pub use availability::QueryableBlock;
 pub use error::Error;
 pub use resolvable::Resolvable;
 
-use data_source::QueryData;
 use futures::Future;
 use hotshot::{certificate, types::SystemContextHandle};
 use hotshot_types::{
@@ -460,13 +390,16 @@ pub struct Options {
 }
 
 /// Run an instance of the HotShot Query service with no customization.
-pub fn run_standalone_service<Types: NodeType, I: NodeImplementation<Types>>(
+pub fn run_standalone_service<Types: NodeType, I: NodeImplementation<Types>, D>(
     _options: &Options,
-    _data_source: QueryData<Types, I, ()>,
+    _data_source: D,
     _hotshot: SystemContextHandle<Types, I>,
 ) -> impl Future<Output = ()> + Send + Sync + 'static
 where
     Block<Types>: QueryableBlock,
+    D: availability::AvailabilityDataSource<Types, I>
+        + status::StatusDataSource
+        + data_source::UpdateDataSource<Types, I>,
 {
     async move { unimplemented!() }
 }
@@ -480,7 +413,10 @@ mod test {
             TransactionHash, TransactionIndex,
         },
         status::{MempoolQueryData, StatusDataSource},
-        testing::mocks::{MockNodeImpl, MockTypes},
+        testing::{
+            consensus::MockDataSource,
+            mocks::{MockNodeImpl, MockTypes},
+        },
     };
     use async_std::{sync::RwLock, task::spawn};
     use async_trait::async_trait;
@@ -499,23 +435,19 @@ mod test {
 
     struct CompositeState {
         store: AtomicStore,
-        hotshot_qs: QueryData<MockTypes, MockNodeImpl, ()>,
+        hotshot_qs: MockDataSource,
         module_state: RollingLog<BincodeLoadStore<u64>>,
     }
 
     #[async_trait]
     impl AvailabilityDataSource<MockTypes, MockNodeImpl> for CompositeState {
-        type LeafStream = <QueryData<MockTypes, MockNodeImpl, ()> as AvailabilityDataSource<
-            MockTypes,
-            MockNodeImpl,
-        >>::LeafStream;
-        type BlockStream = <QueryData<MockTypes, MockNodeImpl, ()> as AvailabilityDataSource<
-            MockTypes,
-            MockNodeImpl,
-        >>::BlockStream;
+        type LeafStream =
+            <MockDataSource as AvailabilityDataSource<MockTypes, MockNodeImpl>>::LeafStream;
+        type BlockStream =
+            <MockDataSource as AvailabilityDataSource<MockTypes, MockNodeImpl>>::BlockStream;
 
         type LeafRange<'a, R> =
-            <QueryData<MockTypes, MockNodeImpl, ()> as AvailabilityDataSource<
+            <MockDataSource as AvailabilityDataSource<
                 MockTypes,
                 MockNodeImpl,
             >>::LeafRange<'a, R>
@@ -523,7 +455,7 @@ mod test {
             Self: 'a,
             R: RangeBounds<usize> + Send;
         type BlockRange<'a, R> =
-            <QueryData<MockTypes, MockNodeImpl, ()> as AvailabilityDataSource<
+            <MockDataSource as AvailabilityDataSource<
                 MockTypes,
                 MockNodeImpl,
             >>::BlockRange<'a, R>
@@ -582,7 +514,7 @@ mod test {
     // Implement data source trait for status API.
     #[async_trait]
     impl StatusDataSource for CompositeState {
-        type Error = <QueryData<MockTypes, MockNodeImpl, ()> as StatusDataSource>::Error;
+        type Error = <MockDataSource as StatusDataSource>::Error;
 
         async fn block_height(&self) -> Result<usize, Self::Error> {
             self.hotshot_qs.block_height().await
@@ -602,7 +534,7 @@ mod test {
     async fn test_composition() {
         let dir = TempDir::new("test_composition").unwrap();
         let mut loader = AtomicStoreLoader::create(dir.path(), "test_composition").unwrap();
-        let hotshot_qs = QueryData::create_with_store(&mut loader, ()).unwrap();
+        let hotshot_qs = MockDataSource::create_with_store(&mut loader, ()).unwrap();
         let module_state =
             RollingLog::create(&mut loader, Default::default(), "module_state", 1024).unwrap();
         let state = CompositeState {
