+++ src/testing/consensus.rs
@@ -10,8 +10,10 @@
 // You should have received a copy of the GNU General Public License along with this program. If not,
 // see <https://www.gnu.org/licenses/>.
 
-use super::mocks::{MockBlock, MockMembership, MockNodeImpl, MockTransaction, MockTypes};
-use crate::{data_source::QueryData, update::UpdateDataSource};
+use super::mocks::{
+    MockBlock, MockMembership, MockNodeImpl, MockTransaction, MockTypes, TestableDataSource,
+};
+use crate::data_source::FileSystemDataSource;
 use async_std::{
     sync::{Arc, RwLock},
     task::spawn,
@@ -34,23 +36,25 @@ use hotshot_types::{
 };
 use std::num::NonZeroUsize;
 use std::time::Duration;
-use tempdir::TempDir;
 
-struct MockNode<UserData> {
-    query_data: Arc<RwLock<QueryData<MockTypes, MockNodeImpl, UserData>>>,
+struct MockNode<D: TestableDataSource> {
     hotshot: SystemContextHandle<MockTypes, MockNodeImpl>,
+    query_data: Arc<RwLock<D>>,
+    _tmp_data: D::TmpData,
 }
 
-pub struct MockNetwork<UserData> {
-    nodes: Vec<MockNode<UserData>>,
-    _dir: TempDir,
+pub struct MockNetwork<D: TestableDataSource> {
+    nodes: Vec<MockNode<D>>,
 }
+
+// MockNetwork can be used with any TestableDataSource, but it's nice to have a default with a
+// convenient type alias.
+pub type MockDataSource = FileSystemDataSource<MockTypes, MockNodeImpl, ()>;
 
 const MINIMUM_NODES: usize = 6;
 
-impl<UserData: Clone + Send> MockNetwork<UserData> {
-    pub async fn init(user_data: UserData) -> Self {
-        let dir = TempDir::new("mock_network").unwrap();
+impl<D: TestableDataSource> MockNetwork<D> {
+    pub async fn init() -> Self {
         let priv_keys = (0..MINIMUM_NODES)
             .map(|_| BN254Priv::generate())
             .collect::<Vec<_>>();
@@ -86,8 +90,6 @@ impl<UserData: Clone + Send> MockNetwork<UserData> {
                 .into_iter()
                 .enumerate()
                 .map(|(node_id, priv_key)| {
-                    let path = dir.path().join(format!("node{}", node_id));
-                    let user_data = user_data.clone();
                     let pub_keys = pub_keys.clone();
                     let known_nodes_with_stake = known_nodes_with_stake.clone();
                     let config = config.clone();
@@ -96,7 +98,7 @@ impl<UserData: Clone + Send> MockNetwork<UserData> {
                         MockMembership::default_election_config(total_nodes.get() as u64);
 
                     async move {
-                        let query_data = QueryData::create(&path, user_data).unwrap();
+                        let (query_data, tmp_data) = D::create(node_id).await;
                         let network = Arc::new(MemoryNetwork::new(
                             pub_keys[node_id],
                             query_data.metrics(),
@@ -131,19 +133,20 @@ impl<UserData: Clone + Send> MockNetwork<UserData> {
                         .await
                         .unwrap();
                         MockNode {
-                            query_data: Arc::new(RwLock::new(query_data)),
                             hotshot,
+                            query_data: Arc::new(RwLock::new(query_data)),
+                            _tmp_data: tmp_data,
                         }
                     }
                 }),
         )
         .await;
 
-        Self { nodes, _dir: dir }
+        Self { nodes }
     }
 }
 
-impl<UserData> MockNetwork<UserData> {
+impl<D: TestableDataSource> MockNetwork<D> {
     pub fn handle(&self) -> SystemContextHandle<MockTypes, MockNodeImpl> {
         self.nodes[0].hotshot.clone()
     }
@@ -152,7 +155,7 @@ impl<UserData> MockNetwork<UserData> {
         self.handle().submit_transaction(tx).await.unwrap();
     }
 
-    pub fn query_data(&self) -> Arc<RwLock<QueryData<MockTypes, MockNodeImpl, UserData>>> {
+    pub fn query_data(&self) -> Arc<RwLock<D>> {
         self.nodes[0].query_data.clone()
     }
 
@@ -167,7 +170,7 @@ impl<UserData> MockNetwork<UserData> {
     }
 }
 
-impl<UserData: Send + Sync + 'static> MockNetwork<UserData> {
+impl<D: TestableDataSource> MockNetwork<D> {
     pub async fn start(&mut self) {
         // Spawn the update tasks.
         for node in &mut self.nodes {
@@ -178,7 +181,7 @@ impl<UserData: Send + Sync + 'static> MockNetwork<UserData> {
                     tracing::info!("EVENT {:?}", event.event);
                     let mut qd = qd.write().await;
                     qd.update(&event).await.unwrap();
-                    qd.commit_version().await.unwrap();
+                    qd.commit_version().await;
                 }
             });
         }
@@ -192,7 +195,7 @@ impl<UserData: Send + Sync + 'static> MockNetwork<UserData> {
     }
 }
 
-impl<UserData> Drop for MockNetwork<UserData> {
+impl<D: TestableDataSource> Drop for MockNetwork<D> {
     fn drop(&mut self) {
         async_std::task::block_on(self.shut_down_impl())
     }
